barberexpress-main/templates/CruzeiroTech/requirements.txt
google.generativeai
python-dotenv
flask
